requirements.txt
flask
PyPDF2
python-docx
Pillow
gunicorn
reportlab
